pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phylosmew"
version = "0.1.0"
description = "Phylogenetic Snakemake evaluation workflow (CLI + Dash)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "GPL-3.0-or-later"}  # adjust
authors = [{ name = "PhyloSmew Authors"}]
dependencies = [
    "typer>=0.9,<1.0",
    "pyyaml>=6",
    "pandas>=1.5",
    "numpy",
    "matplotlib",
    "biopython",
    "dash>=2.17",
    "plotly>=5.20",
    "scikit-learn",
    "statsmodels"
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "mypy"]
# NOTE: Snakemake & tqdist are provided by Conda at install time.

[project.scripts]
phylosmew = "phylosmew.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.package-data]
phylosmew = ["workflow/Snakefile", "workflow/assets/**", "workflow/config.schema.yaml"]
